Link propagation to the main branch
This is a proposal for a feature to permanently redirect all public links to the master -> main branch when the rename tooling is created.  Public repositories, particularly ones that are sources of documentation, are frequently linked externally. Enabling a redirect for these links will make it easier to rename the primary branch.

Thank you for doing this! Very happy to see this tooling coming.